inception

+++ srcs/requirements/wordpress/tools/wordpress.sh
@@ -1,12 +1,31 @@
 #!/bin/sh
 sleep 10
-mv wp-config.php /var/www/html/
-mv wordpress/* /var/www/html/
-#cd /var/www/html;
-#wget http://wordpress.org/latest.tar.gz;
-#tar -xzvf latest.tar.gz;
-#rm -rf latest.tar.gz
-chmod 777 wordpress /var/www/html/wp-content/
-#mv wordpress/* .
-rm -rf wordpress/ latest.tar.gz
-./usr/sbin/php-fpm7 -F;
+
+if [ ! -f /var/www/html/wp-config.php ] ; then
+{
+	wp core download --allow-root
+
+	mv wordpress/* /var/www/html/
+	mv wp-config.php /var/www/html/
+	rm -rf wordpress/ wordpress-6.0-fr_FR.tar.gz
+	#cd /var/www/html/
+
+	echo $MYSQL_USER
+	wp core install		--url="tchappui.42.fr" \
+						--title="ft_containers" \
+						--admin_user=$MYSQL_USER \
+						--admin_password=$MYSQL_PASSWORD \
+						--admin_email=test@gmail.com \
+						--allow-root --path='/var/www/html/' >> log.txt
+
+	wp user create		--allow-root --role=author \
+						$USER2_LOGIN $USER2_MAIL --user_pass=$USER2_PASS \
+						--path='/var/www/html/' >> /log.txt
+
+#	wp plugin update --all --path='/var/www/html/wordpress' >> log.txt
+
+#	wp theme install oceanwp --activate --path='/var/www/html/wordpress' >> log.txt
+}
+fi
+
+./usr/sbin/php-fpm8 -F;
